+++ SpliceMachine.Drda/Requests/ExecutePreparedSqlRequest.cs
@@ -10,7 +10,7 @@ namespace SpliceMachine.Drda
         private readonly Boolean _hasParameters;
 
         public ExecutePreparedSqlRequest(
-            Int32 requestCorrelationId,
+            UInt16 requestCorrelationId,
             UInt16 packageSerialNumber,
             Boolean hasParameters)
         {
@@ -19,7 +19,7 @@ namespace SpliceMachine.Drda
             RequestCorrelationId = requestCorrelationId;
         }
         
-        public Int32 RequestCorrelationId { get; }
+        public UInt16 RequestCorrelationId { get; }
 
         MessageFormat IDrdaRequest.Format => 
             _hasParameters
